ALmost done, all the interactive flow created and just last to finish de README.md, also saved the logo image to use it in the README.md

+++ src/main/java/com/ratelab/utils/PathBuilder.java
@@ -7,7 +7,7 @@ public class PathBuilder {
     static String endpoint = "/pair/";
     static long amount;
 
-    public static String buildPath(String fromCurrency, String toCurrency, int amount) {
+    public static String buildPath(String fromCurrency, String toCurrency, long amount) {
         return baseUrl + apiKey + endpoint + fromCurrency + "/" + toCurrency + "/" + amount;
     }
 }
